pyproject.toml
[project]
name = "test-google-vision-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-auth>=2.40.2",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.16.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
